# backend/postman/openapi.yaml
openapi: "3.0.1"
info:
  title: "Performance Dashboard on AWS"
  description: "Backend REST API"
  version: "2021-02-15T15:50:32Z"
security:
  - cognito-authorizer: []
paths:
...
  /public/dashboard/friendly-url/{friendlyURL}:
    get:
      operationId: GetPublicDashboardByFriendlyURL
      security: []
      parameters:
        - name: friendlyURL
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A public representation of a dashboard including its widgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicDashboard"
  /public/dashboard/{id}:
    get:
      operationId: GetPublicDashboardById
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A public representation of a dashboard including its widgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicDashboard"
...
  /dashboard:
    get:
      operationId: ListDashboards
      responses:
        "200":
          description: List of dashboards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dashboard"

    post:
      operationId: CreateDashboard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDashboardRequest"
      responses:
        "200":
          description: Newly created dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /dashboard/{id}:
    get:
      operationId: GetDashboardById
      parameters:
        - name: id
          in: path
          required: true
          description: Dashboard Id
          schema:
            type: string
      responses:
        "200":
          description: Dashboard with widgets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"

    put:
      operationId: UpdateDashboard
      parameters:
        - name: id
          in: path
          required: true
          description: Dashboard Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDashboardRequest"
      responses:
        "200":
          description: Dashboard updated successfully

    delete:
      operationId: DeleteDashboard
      parameters:
        - name: id
          in: path
          required: true
          description: Dashboard Id
          schema:
            type: string
      responses:
        "200":
          description: Dashboard deleted successfully
  /dashboard/{id}/archive:
    put:
      operationId: ArchiveDashboard
      parameters:
        - name: id
          in: path
          required: true
          description: Dashboard Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveDashboardRequest"
      responses:
        "200":
          description: Dashboard archived successfully
        "409":
          description: Dashboard is not in the expected state
  /dashboard/{id}/publish:
    put:
      operationId: PublishDashboard
      parameters:
        - name: id
          in: path
          required: true
          description: Dashboard Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishDashboardRequest"
      responses:
        "200":
          description: Dashboard published successfully
        "409":
          description: Dashboard is not in the expected state
...
  /topicarea/{id}:
    get:
      operationId: GetTopicareaById
      parameters:
        - name: id
          in: path
          required: true
          description: Topic area Id
          schema:
            $ref: "#/components/schemas/TopicArea"
      responses:
        "200":
          description: topic area name
          content:
            application/json:
              schema:
                type: string

    put:
      operationId: UpdateTopicarea
      parameters:
        - name: id
          in: path
          required: true
          description: Topic area Id
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicAreaRequest"
      responses:
        "200":
          description: Topic area updated successfully

    delete:
      operationId: DeleteTopicarea
      parameters:
        - name: id
          in: path
          required: true
          description: Topic area Id
          schema:
            type: string
      responses:
        "200":
          description: Topic area deleted successfully
...
components:
  securitySchemes:
    cognito-authorizer:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      x-amazon-apigateway-authtype: "cognito_user_pools"

    api_key:
      type: "apiKey"
      name: "x-api-key"
      in: "header"

  # -----------------------------------------
  # Request/Response schemas
  # ------------------------------------------
  schemas:
...
    PublicDashboard:
      type: object
      description: Public representation of a dashboard
      properties:
        id:
          type: string
        name:
          type: string
        topicAreaId:
          type: string
        topicAreaName:
          type: string
        displayTableOfcontents:
          type: boolean
        description:
          type: string
        updatedAt:
          type: string
        friendlyURL:
          type: string
        widgets:
          type: array
          items:
            $ref: "#/components/schemas/Widget"
    Dashboard:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: number
        parentDashboardId:
          type: string
        topicAreaId:
          type: string
        topicAreaName:
          type: string
        displayTableOfContents:
          type: boolean
        description:
          type: string
        createdBy:
          type: string
        updatedAt:
          type: string
        updatedBy:
          type: string
        submittedBy:
          type: string
        publishedBy:
          type: string
        archivedBy:
          type: string
        deletedBy:
          type: string
        state:
          $ref: "#/components/schemas/DashboardState"
        friendlyURL:
          type: string
        releaseNotes:
          type: string
        widgets:
          type: array
          items:
            $ref: "#/components/schemas/Widget"
...
    DashboardState:
      type: string
      enum:
        - Draft
        - Published
        - Archived
        - PublishPending
        - Inactive
...
    Widget:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        widgetType:
          $ref: "#/components/schemas/WidgetType"
        dashboardId:
          type: string
        order:
          type: number
        updatedAt:
          type: string
        showTitle:
          type: boolean
        content:
          type: object
          oneOf:
            - $ref: "#/components/schemas/TextWidgetContent"
            - $ref: "#/components/schemas/ChartWidgetContent"
            - $ref: "#/components/schemas/MetricsWidgetContent"
            - $ref: "#/components/schemas/ImageWidgetContent"
            - $ref: "#/components/schemas/TableWidgetContent"
    WidgetType:
      type: string
      enum:
        - Text
        - Chart
        - Metrics
        - Table
        - Image
    TextWidgetContent:
      type: object
      properties:
        text:
          type: string
    ChartWidgetContent:
      type: object
      properties:
        title:
          type: string
        chartType:
          type: string
          enum:
            - LineChart
            - ColumnChart
            - BarChart
            - PartWholeChart
        datasetId:
          type: string
        summary:
          type: string
        summaryBelow:
          type: boolean
        datasetType:
          type: string
        s3Key:
          $ref: "#/components/schemas/DatasetS3Key"
        fileName:
          type: string
        columnsMetadata:
          $ref: "#/components/schemas/DatasetColumnMetadata"
        sortByColumn:
          type: string
        sortByDesc:
          type: boolean
    MetricsWidgetContent:
      type: object
      properties:
        title:
          type: string
        datasetId:
          type: string
        oneMetricPerRow:
          type: boolean
        datasetType:
          type: string
        s3Key:
          $ref: "#/components/schemas/DatasetS3Key"
    ImageWidgetContent:
      type: object
      properties:
        title:
          type: string
        imageAltText:
          type: string
        summary:
          type: string
        summaryBelow:
          type: boolean
        s3Key:
          $ref: "#/components/schemas/DatasetS3Key"
        fileName:
          type: string
    TableWidgetContent:
      type: object
      properties:
        title:
          type: string
        datasetId:
          type: string
        summary:
          type: string
        summaryBelow:
          type: boolean
        datasetType:
          type: string
        s3Key:
          $ref: "#/components/schemas/DatasetS3Key"
        fileName:
          type: string
        columnsMetadata:
          $ref: "#/components/schemas/DatasetColumnMetadata"
        sortByColumn:
          type: string
        sortByDesc:
          type: boolean
    DatasetColumnMetadata:
      type: object
      properties:
        columnName:
          type: string
        dataType:
          type: string
          enum:
            - Number
            - Text
        hidden:
          type: boolean
    DatasetS3Key:
      type: object
      properties:
        raw:
          type: string
        json:
          type: string
...
    CreateDashboardRequest:
      type: object
      properties:
        topicAreaId:
          type: string
        name:
          type: string
        description:
          type: string
      required:
        - topicAreaId
        - name
    UpdateDashboardRequest:
      type: object
      properties:
        topicAreaId:
          type: string
        name:
          type: string
        description:
          type: string
        updatedAt:
          type: string
        displayTableOfContents:
          type: boolean
      required:
        - topicAreaId
        - name
        - updatedAt
    ArchiveDashboardRequest:
      type: object
      properties:
        updatedAt:
          type: string
      required:
        - updatedAt
    PublishDashboardRequest:
      type: object
      properties:
        updatedAt:
          type: string
        releaseNotes:
          type: string
        friendlyURL:
          type: string
      required:
        - updatedAt
...
    TopicArea:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        createdBy:
          type: string
        dashboardCount:
          type: number
    TopicAreaRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...